requirements.txt
catkin_pkg==1.0.0
colorama==0.4.6
dash==2.18.2
ffmpeg==1.4
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
plotly==5.24.1
PyYAML==6.0.2
PyYAML==6.0.2
ruamel.base==1.0.0
scipy==1.14.1
opencv-python
numba
kaleido
alive_progress
rdp